refactor: drop `test_` prefix from files that don't contain tests
