+++ tests/migration/annotations/remove_migration.py
